# Speed up background map parsing time