Using a benchmark of 110 map zips, 1.8G of data, a couple of changes reduced map parsing time from 20s to 2s.

Performance changes:
- Use ZipFile (requires Java 1.7) instead of ZipFileInputStream  (20s -> 4s)
- Boost the parallelism, introduce multi-threading using half of the number of available cores  (4s -> 2s) 

Commit comment from b3a9d97 with more details about the conversion to ZipFile:

> NewGameChooserModel - Speed up background zip file parsing use ZipFile instead of a ZipFileInputStream. ZipFile can use file system information embedded in the Zip to go directly to the XML files rather than scanning through the entire Zip. As a benchmark, with 110 map zips (1.8G) and this update, map parsing time dropped from 20s to 4s.

Commit comment from 68cf809 with more details about the multi-threading:

> NewGameChooserModel - Multi-threaded background map parsing using half of the available CPU cores. ZipFile already reduced parsing time of 110 maps (1.8G) down to 4s, this change reduces parsing time by an additional 50% to 2s. Analysis on a six core machine showed the sweet spot for the number of threads to be between 2 and 4. Hence the choice of using half the number of cores both falls in this range and is a bit more conservative (for example considering the impact of extra threads on a 4 core or a 2 core machine, +1 and +0 threads is relatively minimal)


<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/triplea-game/triplea/240)
<!-- Reviewable:end -->