@DanVanAtta 2015-10-23
Rebased.. then I fixed the rebase.

There were then some opportunities to simplify the code, so you'll see some additional refactors. The refactors didn't go completely perfectly and you'll see a couple of bug fixes committed at the end to fix those issues.