@ron-murhammer 2015-11-09
@DanVanAtta This seems to work well and is much faster for me now. I think the major win is switching to ZipFile instead of ZipInputStream as it appears to be around 5-10 times faster. Merging this now.